server/pyproject.toml
[tool.poetry]
package-mode = false


[tool.poetry.dependencies]
python = "^3.10"
django-environ = "^0.11.2"
django = "^5.1.4"
pillow = "^11.0.0"
aiohttp = "^3.11.10"
daphne = "^4.1.2"
channels = "^4.2.0"
channels-redis = "^4.2.1"
psycopg2-binary = "^2.9.10"
django-redis = "^5.4.0"
djangorestframework = "^3.15.2"
markdown = "^3.7"
django-filter = "^24.3"
celery = "^5.4.0"
redis = "^5.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
